Obliterated the onos-topo/api/device; the old device API
